Create linter NoLiteralArguments
Instead of automatically fixing the problem, we can report an error
message to complain about literals that appear in argument position.
The benefit of writing this analysis as a linter is that it can be
syntactic since we don't need to know the parameter name.

Linters don't need an output test file since they only report error
message. To assert that an error message is reported at a particular
position, add `// assert: NoLiteralArguments` comments to the
lines where you expect the diagnostic to be reported. Optionally,
you can assert the message and position of the diagnostic by using
multi-line `/*` comments
```
  complete(false) /* assert: NoLiteralArguments
           ^^^^^
  Use named arguments for literals such as 'parameterName = false'
  */
```